Add `AppsLightboxImageStore`
This component stores image data for the apps lightbox in one place.
