+++ dotcom-rendering/src/components/AppsLightboxImage.importable.tsx
@@ -1,14 +1,12 @@
 import { css } from '@emotion/react';
-import { Image } from '@guardian/bridget/Image';
-import { getGalleryClient } from '../lib/bridgetApi';
-import { generateImageURL } from '../lib/image';
-import type { ImageForAppsLightbox } from '../model/appsLightboxImages';
 import { type RoleType } from '../types/content';
 import { Picture } from './Picture';
+import { getOnClickHandler } from './Foo.importable';
+import { element } from 'screenfull';
 
 type Props = {
-	images: ImageForAppsLightbox[];
 	currentIndex: number;
+	elementId: string;
 	role: RoleType;
 	format: ArticleFormat;
 	master: string;
@@ -20,8 +18,8 @@ type Props = {
 };
 
 export const AppsLightboxImage = ({
-	images,
 	currentIndex,
+	elementId,
 	role,
 	format,
 	master,
@@ -32,32 +30,6 @@ export const AppsLightboxImage = ({
 	isLazy = true,
 }: Props) => {
 	const hasLightbox = currentIndex !== -1;
-	const onClick = () => {
-		// Handle the case the device is rotated
-		const imageWidth = Math.max(window.innerHeight, window.innerWidth);
-		const resolution = window.devicePixelRatio >= 2 ? 'high' : 'low';
-		void getGalleryClient()
-			.launchSlideshow(
-				images.map(
-					(image) =>
-						new Image({
-							width: image.width,
-							height: image.height,
-							url: generateImageURL({
-								mainImage: image.masterUrl,
-								imageWidth,
-								resolution,
-							}),
-							caption: image.caption,
-							credit: image.credit,
-						}),
-				),
-				currentIndex,
-				document.title,
-			)
-			// we don't need to handle this error
-			.catch(() => undefined);
-	};
 	const picture = (
 		<Picture
 			role={role}
@@ -72,7 +44,7 @@ export const AppsLightboxImage = ({
 	);
 	return hasLightbox ? (
 		<button
-			onClick={onClick}
+			onClick={getOnClickHandler(elementId)}
 			type="button"
 			css={css`
 				border: none;
