+++ dotcom-rendering/src/components/ImageComponent.tsx
@@ -10,7 +10,6 @@ import {
 import { decidePalette } from '../lib/decidePalette';
 import { getLargest, getMaster } from '../lib/image';
 import { isWideEnough } from '../lib/lightbox';
-import type { ImageForAppsLightbox } from '../model/appsLightboxImages';
 import { palette as themePalette } from '../palette';
 import type { Switches } from '../types/config';
 import type { ImageBlockElement, RoleType } from '../types/content';
@@ -34,7 +33,6 @@ type Props = {
 	title?: string;
 	isAvatar?: boolean;
 	switches?: Switches;
-	imagesForAppsLightbox: ImageForAppsLightbox[];
 };
 
 const starsWrapper = css`
@@ -241,7 +239,6 @@ export const ImageComponent = ({
 	title,
 	isAvatar,
 	switches,
-	imagesForAppsLightbox,
 }: Props) => {
 	const { renderingTarget } = useConfig();
 	// Its possible the tools wont send us any images urls
@@ -321,10 +318,7 @@ export const ImageComponent = ({
 				{renderingTarget === 'Apps' ? (
 					<Island priority="critical">
 						<AppsLightboxImage
-							images={imagesForAppsLightbox}
-							currentIndex={imagesForAppsLightbox.findIndex(
-								(img) => img.elementId === element.elementId,
-							)}
+							elementId={element.elementId}
 							role={role}
 							format={format}
 							master={image.url}
@@ -388,10 +382,7 @@ export const ImageComponent = ({
 				{renderingTarget === 'Apps' ? (
 					<Island priority="critical">
 						<AppsLightboxImage
-							images={imagesForAppsLightbox}
-							currentIndex={imagesForAppsLightbox.findIndex(
-								(img) => img.elementId === element.elementId,
-							)}
+							elementId={element.elementId}
 							role={role}
 							format={format}
 							master={image.url}
@@ -458,10 +449,7 @@ export const ImageComponent = ({
 				{renderingTarget === 'Apps' ? (
 					<Island priority="critical">
 						<AppsLightboxImage
-							images={imagesForAppsLightbox}
-							currentIndex={imagesForAppsLightbox.findIndex(
-								(img) => img.elementId === element.elementId,
-							)}
+							elementId={element.elementId}
 							role={role}
 							format={format}
 							master={image.url}
