+++ dotcom-rendering/src/layouts/StandardLayout.tsx
@@ -58,6 +58,7 @@ import { palette as themePalette } from '../palette';
 import type { DCRArticle } from '../types/frontend';
 import type { RenderingTarget } from '../types/renderingTarget';
 import { BannerWrapper, Stuck } from './lib/stickiness';
+import { Foo } from '../components/Foo.importable';
 
 const StandardGrid = ({
 	children,
@@ -491,6 +492,11 @@ export const StandardLayout = (props: WebProps | AppProps) => {
 				{isApps && (
 					<Island priority="critical">
 						<AdPortals />
+					</Island>
+				)}
+				{isApps && (
+					<Island priority="critical">
+						<Foo images={article.imagesForAppsLightbox} />
 					</Island>
 				)}
 				<Section
